+++ lib/campaign.js
@@ -3,7 +3,7 @@
 let _ = require('lodash')
   , parseOneAddress = require('email-addresses').parseOneAddress
   , version = require('../package.json').version
-  , translator = require('../lib/translate')
+  , translator = require('../lib/translator')
   , rules = [
   [/\[(Sender_[A-Za-z0-0]+)]/g, "{{ $1 or '' }}"],
   //[/\[Sender_Name]/g, "{{ Sender_Name or '' }}"],
