+++ lib/transactional.js
@@ -1,6 +1,6 @@
 'use strict';
 let _ = require('lodash')
-  , translator = require('../lib/translate')
+  , translator = require('../lib/translator')
   , version = require('../package.json').version,
   rules = [
     [/<%subject%>/, '{{ subject }}'],
