+++ lib/translator.js
